requirements.txt
fastapi
uvicorn
jinja2
python-multipart
transformers
torch
bitsandbytes
pillow
accelerate